requirements.txt
pyserial
# serial port access setting
# ex) sudo chmod a+rw /dev/ttyACM0  
#